simulator/pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".

[tool.poetry]
name = "simulator"
version = "0.0.0"
description = ""
authors = [ "Autonomous Llama <paint-llama@gmail.com>" ]
readme = "README.md"

  [tool.poetry.dependencies]
  pycairo = "^1.27.0"
  pydantic-settings = "^2.6.1"
  pydantic = "^2.9.2"
  python = "3.12.7"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
